backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-auth-ldap==5.2.0
django-stubs==5.2.0
django-stubs-ext==5.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-ldap==3.4.4
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.13.2
